+++ index.html
@@ -40,6 +40,7 @@
         </div>
       </div>
     </div>
+    <script type="module" src="js/constants.js" defer></script>
     <script type="module" src="js/addToTask.js" defer></script>
   </body>
 </html>
