+++ js/constants.js
@@ -1,0 +1,15 @@
+//create button
+export const createTask = document.getElementById("createButton");
+
+//content  task list
+export const content = document.querySelector(".content");
+export const list = document.getElementById("list");
+
+//Input BOX: add to task input box, add task button, trash button
+export const inputTask = document.getElementById("inputTask");
+export const input = document.getElementById("input");
+export const addTask = document.getElementById("addTaskButton");
+export const trashInput = document.querySelector(".trashInput");
+
+//required variable
+export const today = new Date();
